tests/submission/Infraware/Forms/contents/Forms/Not Use/option_disabled.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>option_disabled</h3>
	</p>
	
	<li>The second option control be disabled when the document was loaded.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3" selected> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var option = select.item(1);
		option.disabled = true;

		if (typeof(option.disabled) == "boolean") {
			if (option.disabled != true) {
				test(function() {
					assert_true(false, "disabled attribute is not exist.");
				});
			}
		} else {
			test(function() {
				assert_unreached("disabled attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
